+++ src/common/components/Icon/icons/itemList.jsx
@@ -1,0 +1,32 @@
+import Svg from '../svg';
+
+const itemList = ({ className }) => (
+  <Svg className={className} originalSize={24}>
+    <g clipPath="url(#clip0_1282_40268)" id="Frame">
+      <path
+        d="M7 12.46C7 13.786 7.52678 15.0578 8.46447 15.9955C9.40215 16.9332 10.6739 17.46 12 17.46C13.3261 17.46 14.5979 16.9332 15.5355 15.9955C16.4732 15.0578 17 13.786 17 12.46C17 11.1339 16.4732 9.86211 15.5355 8.92443C14.5979 7.98675 13.3261 7.45996 12 7.45996C10.6739 7.45996 9.40215 7.98675 8.46447 8.92443C7.52678 9.86211 7 11.1339 7 12.46Z"
+        fill="none"
+        id="Vector"
+        stroke="#00A9E0"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        strokeWidth="2"
+      />
+      <path
+        d="M11 12.46C11 12.7252 11.1054 12.9795 11.2929 13.1671C11.4804 13.3546 11.7348 13.46 12 13.46C12.2652 13.46 12.5196 13.3546 12.7071 13.1671C12.8946 12.9795 13 12.7252 13 12.46C13 12.1947 12.8946 11.9404 12.7071 11.7529C12.5196 11.5653 12.2652 11.46 12 11.46C11.7348 11.46 11.4804 11.5653 11.2929 11.7529C11.1054 11.9404 11 12.1947 11 12.46Z"
+        id="Vector_2"
+        stroke="#00A9E0"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        strokeWidth="2"
+      />
+    </g>
+    <defs>
+      <clipPath id="clip0_1282_40268">
+        <rect fill="white" height="24" transform="translate(0 0.820312)" width="24" />
+      </clipPath>
+    </defs>
+  </Svg>
+);
+
+export { itemList };